+++ analytics/src/main/java/com/bluetriangle/analytics/dynamicconfig/model/BTTRemoteConfiguration.kt
@@ -13,7 +13,14 @@ internal open class BTTRemoteConfiguration(
     val enableScreenTracking: Boolean,
     val enableGrouping: Boolean,
     val groupingIdleTime: Int,
-    val groupedViewSampleRate: Double?
+    val groupedViewSampleRate: Double?,
+    val enableGroupingTapDetection: Boolean,
+    val enableNetworkStateTracking: Boolean,
+    val enableCrashTracking: Boolean,
+    val enableANRTracking: Boolean,
+    val enableMemoryWarning: Boolean,
+    val enableLaunchTime: Boolean,
+    val enableWebViewStitching: Boolean
 ) {
     override fun equals(other: Any?): Boolean {
         if(other is BTTRemoteConfiguration) {
@@ -24,13 +31,20 @@ internal open class BTTRemoteConfiguration(
                     other.enableScreenTracking == enableScreenTracking &&
                     other.enableGrouping == enableGrouping &&
                     other.groupingIdleTime == groupingIdleTime &&
-                    other.groupedViewSampleRate == groupedViewSampleRate
+                    other.groupedViewSampleRate == groupedViewSampleRate &&
+                    other.enableGroupingTapDetection == enableGroupingTapDetection &&
+                    other.enableNetworkStateTracking == enableNetworkStateTracking &&
+                    other.enableCrashTracking == enableCrashTracking &&
+                    other.enableANRTracking == enableANRTracking &&
+                    other.enableMemoryWarning == enableMemoryWarning &&
+                    other.enableLaunchTime == enableLaunchTime &&
+                    other.enableWebViewStitching == enableWebViewStitching
         }
         return false
     }
 
     override fun toString(): String {
-        return "RemoteConfig { networkSampleRate: $networkSampleRate, ignoreList: ${ignoreScreens}, enableRemoteConfigAck: $enableRemoteConfigAck, enableAllTracking: $enableAllTracking,  enableScreenTracking: $enableScreenTracking, enableGrouping: $enableGrouping, groupingIdleTime: $groupingIdleTime, groupedViewSampleRate: $groupedViewSampleRate }"
+        return "RemoteConfig { networkSampleRate: $networkSampleRate, ignoreList: ${ignoreScreens}, enableRemoteConfigAck: $enableRemoteConfigAck, enableAllTracking: $enableAllTracking,  enableScreenTracking: $enableScreenTracking, enableGrouping: $enableGrouping, groupingIdleTime: $groupingIdleTime, groupedViewSampleRate: $groupedViewSampleRate, enableGroupingTapDetection: $enableGroupingTapDetection, enableNetworkStateTracking: $enableNetworkStateTracking, enableCrashTracking: $enableCrashTracking, enableANRTracking: $enableANRTracking, enableMemoryWarning: $enableMemoryWarning, enableLaunchTime: $enableLaunchTime, enableWebViewStitching: $enableWebViewStitching }"
     }
 
     override fun hashCode(): Int {
@@ -42,6 +56,13 @@ internal open class BTTRemoteConfiguration(
         result = 31 * result + groupingIdleTime
         result = 31 * result + ignoreScreens.hashCode()
         result = 31 * result + groupedViewSampleRate.hashCode()
+        result = 31 * result + enableGroupingTapDetection.hashCode()
+        result = 31 * result + enableNetworkStateTracking.hashCode()
+        result = 31 * result + enableCrashTracking.hashCode()
+        result = 31 * result + enableANRTracking.hashCode()
+        result = 31 * result + enableMemoryWarning.hashCode()
+        result = 31 * result + enableLaunchTime.hashCode()
+        result = 31 * result + enableWebViewStitching.hashCode()
         return result
     }