+++ analytics/src/main/java/com/bluetriangle/analytics/dynamicconfig/model/BTTSavedRemoteConfigurationMapper.kt
@@ -6,7 +6,14 @@
 package com.bluetriangle.analytics.dynamicconfig.model
 
 import com.bluetriangle.analytics.Constants
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_ANR_TRACKING
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_CRASH_TRACKING
 import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_GROUPING
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_LAUNCH_TIME
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_MEMORY_WARNING
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_NETWORK_STATE_TRACKING
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_WEB_VIEW_STITCHING
+import com.bluetriangle.analytics.Constants.DEFAULT_ENABLE_GROUPING_TAP_DETECTION
 import com.bluetriangle.analytics.utility.getBooleanOrNull
 import com.bluetriangle.analytics.utility.getDoubleOrNull
 import com.bluetriangle.analytics.utility.getIntOrNull
@@ -24,6 +31,13 @@ internal object BTTSavedRemoteConfigurationMapper {
     private const val GROUPING_IDLE_TIME = "groupingIdleTime"
     private const val ENABLE_SCREEN_TRACKING = "enableScreenTracking"
     private const val GROUPED_VIEW_SAMPLE_RATE = "groupedViewSampleRate"
+    private const val ENABLE_GROUPING_TAP_DETECTION = "enableGroupingTapDetection"
+    private const val ENABLE_NETWORK_STATE_TRACKING = "enableNetworkStateTracking"
+    private const val ENABLE_CRASH_TRACKING = "enableCrashTracking"
+    private const val ENABLE_ANR_TRACKING = "enableANRTracking"
+    private const val ENABLE_MEMORY_WARNING = "enableMemoryWarning"
+    private const val ENABLE_LAUNCH_TIME = "enableLaunchTime"
+    private const val ENABLE_WEB_VIEW_STITCHING = "enableWebViewStitching"
 
     fun fromJson(jsonObject: JSONObject, defaultConfig: BTTRemoteConfiguration): BTTSavedRemoteConfiguration {
         val ignoreScreens = jsonObject.getJsonArrayOrNull(IGNORE_SCREENS)?.let { array ->
@@ -45,6 +59,13 @@ internal object BTTSavedRemoteConfigurationMapper {
             jsonObject.getBooleanOrNull(ENABLE_GROUPING)?:DEFAULT_ENABLE_GROUPING,
             jsonObject.getIntOrNull(GROUPING_IDLE_TIME)?: Constants.DEFAULT_GROUPING_IDLE_TIME,
             jsonObject.getDoubleOrNull(GROUPED_VIEW_SAMPLE_RATE),
+            jsonObject.getBooleanOrNull(ENABLE_GROUPING_TAP_DETECTION) ?: DEFAULT_ENABLE_GROUPING_TAP_DETECTION,
+            jsonObject.getBooleanOrNull(ENABLE_NETWORK_STATE_TRACKING) ?: DEFAULT_ENABLE_NETWORK_STATE_TRACKING,
+            jsonObject.getBooleanOrNull(ENABLE_CRASH_TRACKING) ?: DEFAULT_ENABLE_CRASH_TRACKING,
+            jsonObject.getBooleanOrNull(ENABLE_ANR_TRACKING) ?: DEFAULT_ENABLE_ANR_TRACKING,
+            jsonObject.getBooleanOrNull(ENABLE_MEMORY_WARNING) ?: DEFAULT_ENABLE_MEMORY_WARNING,
+            jsonObject.getBooleanOrNull(ENABLE_LAUNCH_TIME) ?: DEFAULT_ENABLE_LAUNCH_TIME,
+            jsonObject.getBooleanOrNull(ENABLE_WEB_VIEW_STITCHING) ?: DEFAULT_ENABLE_WEB_VIEW_STITCHING,
             jsonObject.getLong(SAVED_DATE)
         )
     }
@@ -62,6 +83,13 @@ internal object BTTSavedRemoteConfigurationMapper {
         put(ENABLE_GROUPING, config.enableGrouping)
         put(GROUPING_IDLE_TIME, config.groupingIdleTime)
         put(GROUPED_VIEW_SAMPLE_RATE, config.groupedViewSampleRate)
+        put(ENABLE_GROUPING_TAP_DETECTION, config.enableGroupingTapDetection)
+        put(ENABLE_NETWORK_STATE_TRACKING, config.enableNetworkStateTracking)
+        put(ENABLE_CRASH_TRACKING, config.enableCrashTracking)
+        put(ENABLE_ANR_TRACKING, config.enableANRTracking)
+        put(ENABLE_MEMORY_WARNING, config.enableMemoryWarning)
+        put(ENABLE_LAUNCH_TIME, config.enableLaunchTime)
+        put(ENABLE_WEB_VIEW_STITCHING, config.enableWebViewStitching)
         put(SAVED_DATE, config.savedDate)
     }